Split up index into individual files in components folder
